backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
surya-ocr
opencv-python
numpy
Pillow
pyspellchecker
pypdfium2
